% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.Norm.R
\name{batch.Norm}
\alias{batch.Norm}
\title{Implement the ComBat normalization method developed for microarrays via the sva package.
This corrects for batch effects related to the slide (Sentrix_ID)
Uses an empirical Bayes method to correct for technical variation (after logit transformation, required as beta values are 0-1, and followed by reverse logit transformation)
Will abort if all samples are from a single slide}
\usage{
batch.Norm(batchNorm_name = "batchNorm_ComBat_normalized_EPIC_betaVals")
}
\arguments{
\item{myNorm_name}{character(1) The name of the output file returned from the champ.norm function.}
}
\value{
A loaded object containing ComBat normalized beta values.
}
\description{
Implement the ComBat normalization method developed for microarrays via the sva package.
This corrects for batch effects related to the slide (Sentrix_ID)
Uses an empirical Bayes method to correct for technical variation (after logit transformation, required as beta values are 0-1, and followed by reverse logit transformation)
Will abort if all samples are from a single slide
}
\details{
This approach uses the ChAMP package first described by Morris TJ et al:
Morris TJ, Butcher LM, Teschendorff AE, Chakravarthy AR, Wojdacz TK and Beck S (2014). “ChAMP: 450k Chip Analysis Methylation
Pipeline.” _Bioinformatics_, *30*(3), pp. 428-430. doi: 10.1093/bioinformatics/btt684
(URL:http://doi.org/10.1093/bioinformatics/btt684).
}

