% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.thresh.R
\name{beta.thresh}
\alias{beta.thresh}
\title{If input file name is found in the working directory, the file is loaded directly}
\usage{
beta.thresh(limma.comparison, deltaB)
}
\arguments{
\item{limma.comparison}{data.frame(1) A data frame returned from the limma.MVP function}

\item{deltaB}{numeric(1) A numeric value equal to the fraction}
}
\value{
A table of
}
\description{
If input file name is found in the working directory, the file is loaded directly
}
\details{
This approach uses the ChAMP package first described by Morris TJ et al:
Morris TJ, Butcher LM, Teschendorff AE, Chakravarthy AR, Wojdacz TK and Beck S (2014). “ChAMP: 450k Chip Analysis Methylation
Pipeline.” _Bioinformatics_, *30*(3), pp. 428-430. doi: 10.1093/bioinformatics/btt684
(URL:http://doi.org/10.1093/bioinformatics/btt684).
}

