% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmiq.Norm.R
\name{bmiq.Norm}
\alias{bmiq.Norm}
\title{Performs beta-mixture quantile (BMIQ) normalization using the champ.norm() function
BMIQ is an intra-sample normalization procedure, correcting for type-2 probe values (methylation arrays have two types of probes that yield different distributions).
3-step procedure: (i) fitting of a 3-state beta mixture model, (ii) transformation of state-membership probabilities of type2 probes
into quantiles of the type1 distribution, and (iii) a conformal transformation for the hemi-methylated probes.#' Normalize Imported IDAT files using 'champ.norm'}
\usage{
bmiq.Norm(myNorm_name = "myNorm_BMIQ_normalized_EPIC_betaVals")
}
\arguments{
\item{myNorm_name}{character(1) The name of the output file returned from the champ.norm function.}
}
\value{
A loaded object containing BMIQ normalized beta values.
}
\description{
Performs beta-mixture quantile (BMIQ) normalization using the champ.norm() function
BMIQ is an intra-sample normalization procedure, correcting for type-2 probe values (methylation arrays have two types of probes that yield different distributions).
3-step procedure: (i) fitting of a 3-state beta mixture model, (ii) transformation of state-membership probabilities of type2 probes
into quantiles of the type1 distribution, and (iii) a conformal transformation for the hemi-methylated probes.#' Normalize Imported IDAT files using 'champ.norm'
}
\details{
This approach uses the ChAMP package first described by Morris TJ et al:
Morris TJ, Butcher LM, Teschendorff AE, Chakravarthy AR, Wojdacz TK and Beck S (2014). “ChAMP: 450k Chip Analysis Methylation
Pipeline.” _Bioinformatics_, *30*(3), pp. 428-430. doi: 10.1093/bioinformatics/btt684
(URL:http://doi.org/10.1093/bioinformatics/btt684).
}

