% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGI.load.R
\name{CGI.load}
\alias{CGI.load}
\title{Implements the 'CGI.load' function to import UCSC CpG island (hg19) annotations}
\usage{
CGI.load(file_path = getwd(), CGI_file.pre = "cpgIslandExt",
  CGI_file.post = "UCSC_hg19_cpgIsland")
}
\arguments{
\item{file_path}{character(1) The name of the directory containing the UCSC CGI annotation file.}

\item{CGI_file.pre}{character(1) The name of the original UCSC CGI annotation file.}

\item{CGI_file.post}{character(1) The name of the saved UCSC CGI file.}
}
\value{
A data frame containing UCSC CpG island (hg19).
}
\description{
Implements the 'CGI.load' function to import UCSC CpG island (hg19) annotations
}
\details{
The CGI.load() function is used to import and organize the UCSC CpG island annotation files (hg19/grch37)
         further analyses. The function looks for the original file, 'cpgIslandExt.txt' (by default, set by the CGI_file argument),
         in the default directory (set by the file_path argument). It filters information for only CGIs
         on chromosomes 1-22, X, Y & MT. The file is sorted by location and resaved as "refGene_hg19_info.txt" (by default,
         set the refSeq.post argument). If this file has been saved, the function will directly load it instead.
}

