% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl.load.R
\name{ensembl.load}
\alias{ensembl.load}
\title{Implements the 'ensembl.load' function to import ensembl (hg19) gene annotations}
\usage{
ensembl.load(file_path = getwd(), file_name = "grch37_geneInfo.txt",
  ensembl_attributes = c("ensembl_gene_id", "ensembl_transcript_id",
  "chromosome_name", "start_position", "end_position", "strand",
  "transcript_start", "transcript_end", "transcription_start_site",
  "transcript_length", "external_gene_name", "percentage_gc_content"))
}
\arguments{
\item{file_path}{character(1) The name of the directory containing the refSeq annotation file.}

\item{file_name}{character(1) The name of the saved ensembl annotation file.}

\item{ensembl_attributes}{character(1) The name of the ensembl attributes imported through biomaRt.}
}
\value{
A data frame containing ensembl (hg19) annotations with TSS information
}
\description{
Implements the 'ensembl.load' function to import ensembl (hg19) gene annotations
}
\details{
The ensembl.load() function is used to import and organize the ensembl annotation files (hg19/grch37) using
         biomaRt for further processing. Specifically, TSS location is determined based on strand information and TxStart/TxEnd.
         The function looks for the original file, 'refGene_nochr.txt' (by default, set by the refSeq.pre argument),
         in the current directory (by default, set bu the file_path argument). It filters information for only genes
         on chromosomes 1-22, X, Y & MT, then identifies TSS locations as described above. The file is sorted by
         location and resaved as "refGene_hg19_info.txt" (by default, set the refSeq.post argument). If this file
         has been saved, the function will directly load it instead.
}

