% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idat.Load.R
\name{idat.Load}
\alias{idat.Load}
\title{Import IDAT files using 'champ.load'}
\usage{
idat.Load(IDAT_path, myLoad_name = "currentStudyloadedData")
}
\arguments{
\item{IDAT_path}{character(1) The path to the IDAT data files.}

\item{myLoad_name}{character(1) The name of the output file returned from the champ.load function.}
}
\value{
A loaded object (list) containing imported IDAT information for further processing.
}
\description{
Import IDAT files using 'champ.load'
}
\details{
This approach uses the ChAMP package first described by Morris TJ et al:
Morris TJ, Butcher LM, Teschendorff AE, Chakravarthy AR, Wojdacz TK and Beck S (2014). “ChAMP: 450k Chip Analysis Methylation
Pipeline.” _Bioinformatics_, *30*(3), pp. 428-430. doi: 10.1093/bioinformatics/btt684
(URL:http://doi.org/10.1093/bioinformatics/btt684).
}

