% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refSeq.load.R
\name{refSeq.load}
\alias{refSeq.load}
\title{Implements the 'refSeq.load' function to import refSeq (hg19) gene annotations}
\usage{
refSeq.load(file_path = getwd(), refSeq.pre = "refGene_nochr.txt",
  refSeq.post = "refGene_hg19_info.txt")
}
\arguments{
\item{file_path}{character(1) The name of the directory containing the refSeq annotation file.}

\item{refSeq.pre}{character(1) The name of the file containing the original refSeq annotation file.}

\item{refSeq.post}{character(1) The name of the file containing the organized/saved refSeq
annotation file with TSS information.}
}
\value{
A data frame containing refSeq (hg19) annotations with corrected TSS information
}
\description{
Implements the 'refSeq.load' function to import refSeq (hg19) gene annotations
}
\details{
The refseq.load() function is used to import and organize the refSeq annotation file (hg19) for further
         processing. Specifically, TSS location is determined based on strand information and TxStart/TxEnd.
         The function looks for the original file, 'refGene_nochr.txt' (by default, set by the refSeq.pre argument),
         in the current directory (by default, set bu the file_path argument). It filters information for only genes
         on chromosomes 1-22, X, Y & MT, then identifies TSS locations as described above. The file is sorted by
         location and resaved as "refGene_hg19_info.txt" (by default, set the refSeq.post argument). If this file
         has been saved, the function will directly load it instead.
}

